+++ assets/stylesheets/main.css
@@ -2,7 +2,7 @@ input {
   padding: 7px;
   height: 40px;
   width: 70%;
-   margin-left: 15%; 
+  margin-left: 15%;
 }
 
 .f2 {
@@ -20,6 +20,7 @@ form {
   padding: 10px 0 10px 0;
   border-radius: 20px;
 }
+
 .b-1 {
   margin-left: 45%;
   align-self: center;
